+++ tests/spec/core_functions/color/hsl/four_args/clamped.rs
@@ -14,16 +14,16 @@ mod alpha {
         assert_eq!(
             runner().ok("a {b: hsl(0, 100%, 50%, 1.1)}\n"),
             "a {\
-         \n  b: red;\
+         \n  b: hsl(0deg, 100%, 50%);\
          \n}\n"
         );
     }
     #[test]
     fn below() {
         assert_eq!(
-            runner().ok("a {b: rgba(0, 100%, 50%, -0.1)}\n"),
+            runner().ok("a {b: hsla(0, 100%, 50%, -0.1)}\n"),
             "a {\
-         \n  b: rgba(0, 255, 128, 0);\
+         \n  b: hsla(0deg, 100%, 50%, 0);\
          \n}\n"
         );
     }
@@ -33,7 +33,7 @@ fn blue() {
     assert_eq!(
         runner().ok("a {b: hsl(0, 100%, 9999%, 0.5)}\n"),
         "a {\
-         \n  b: rgba(255, 255, 255, 0.5);\
+         \n  b: hsla(0deg, 100%, 100%, 0.5);\
          \n}\n"
     );
 }
@@ -42,7 +42,7 @@ fn saturation() {
     assert_eq!(
         runner().ok("a {b: hsl(0, -0.1%, 50%, 0.5)}\n"),
         "a {\
-         \n  b: rgba(128, 128, 128, 0.5);\
+         \n  b: hsla(0deg, 0%, 50%, 0.5);\
          \n}\n"
     );
 }