+++ tests/spec/core_functions/color/rgb/four_args/alpha.rs
@@ -18,7 +18,7 @@ mod percent {
             assert_eq!(
                 runner().ok("a {b: rgb(0, 0, 0, 250%)}\n"),
                 "a {\
-         \n  b: black;\
+         \n  b: rgb(0, 0, 0);\
          \n}\n"
             );
         }
@@ -28,7 +28,7 @@ mod percent {
         assert_eq!(
             runner().ok("a {b: rgb(0, 0, 0, 100%)}\n"),
             "a {\
-         \n  b: black;\
+         \n  b: rgb(0, 0, 0);\
          \n}\n"
         );
     }
@@ -73,7 +73,7 @@ mod unitless {
             assert_eq!(
                 runner().ok("a {b: rgb(0, 0, 0, 250)}\n"),
                 "a {\
-         \n  b: black;\
+         \n  b: rgb(0, 0, 0);\
          \n}\n"
             );
         }
@@ -83,7 +83,7 @@ mod unitless {
         assert_eq!(
             runner().ok("a {b: rgb(0, 0, 0, 1)}\n"),
             "a {\
-         \n  b: black;\
+         \n  b: rgb(0, 0, 0);\
          \n}\n"
         );
     }